confirm modal correct for delete

+++ src/app/modals/modal-session-details/modal-session-details.component.ts
@@ -8,6 +8,8 @@ import { Session } from '../../interfaces/session.interface';
 import { SessionService } from '../../services/session/session.service';
 import { ModalSessionListComponent } from '../modal-session-list/modal-session-list.component';
 import { ToastrModule, ToastrService } from 'ngx-toastr';
+import { ConfirmService } from '../../services/confirm.service';
+import { ConfirmModalComponent } from '../confirm-modal/confirm-modal.component';
 
 @Component({
   selector: 'app-modal-session-details',
@@ -22,13 +24,17 @@ export class ModalSessionDetailsComponent {
   private toastrService = inject(ToastrService);
 
   private sessionService = inject(SessionService);
+  private confirmService = inject(ConfirmService);
 
   public selectedSessionId!: number | null;
   public sessionData!: Observable<Session[] | null>;
-  public selectedStudentId!:number | null;
+  public selectedStudentId!: number | null;
+  public confirmDeleteElement: Observable<boolean>;
+
 
   constructor() {
     this.sessionData = this.sessionService.sessionData;
+    this.confirmDeleteElement = this.confirmService.confirmDeleteElement;
   }
 
   ngOnInit(): void {
@@ -55,22 +61,36 @@ export class ModalSessionDetailsComponent {
     modalRef.componentInstance.selectedStudentId = this.selectedStudentId;
   }
 
-  public closeModal(){
+  public deleteSession() {
+    this.openConfirmModal();
+    this.confirmDeleteElement.subscribe((value: boolean) => {
+      if (value) {
+        this.sessionService.deleteSession(this.selectedSessionId).subscribe({
+          next: (response) => {
+            this, this.showSuccess();
+            this.modalService.dismissAll();
+          },
+          error: (error) => {
+            this.showError();
+            console.error('Error al eliminar la sesión', error)
+          },
+        });
+      }
+    });
+  }
+
+  public openConfirmModal() {
+    const modalRef = this.modalService.open(ConfirmModalComponent, {
+      centered: true,
+      backdrop: 'static',
+    });
+    (modalRef.componentInstance as ConfirmModalComponent).setModalRef(modalRef);
+  }
+
+  public closeModal() {
     this.modalService.dismissAll();
   }
 
-  public deleteSession(){
-    this.sessionService.deleteSession(this.selectedSessionId).subscribe({
-      next: (response) => {
-        this,this.showSuccess(); //Todo: need the confirm modal
-        this.modalService.dismissAll();
-      },
-      error: (error) => {
-        this.showError();
-        console.error('Error al eliminar la sesión', error)
-      },
-    });
-  }
 
   private showSuccess() {
     this.toastrService.success('Los sesión se ha eliminado correctamente!', 'Felicidades!');
